+++ analysis/tests/resources/dummy_tool.ps1
@@ -1,7 +1,9 @@
 Write-Output "Dummy tool!"
-Start-Sleep -Seconds 5
-Write-Output ("Your argument: " + $args[0])
-Write-Output ("SUT: " + $args[1])
+try {
+    Start-Sleep -Seconds $args[0] > $null
+} catch { }
+Write-Output ("Your argument: " + $args[1])
+Write-Output ("SUT: " + $args[2])
 Set-Content -Path ./file1 -Value 'test file'
 mkdir -Force dir1 > $null
 Set-Content -Path ./dir1/file2 -Value 'test file'