+++ compilation/src/main/java/cz/vutbr/fit/group/verifit/oslc/compilation/automationRequestExecution/sutFetcher/SutFetchGit.java
@@ -19,6 +19,7 @@ import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.api.errors.JGitInternalException;
 
+import cz.vutbr.fit.group.verifit.oslc.OslcValues;
 import cz.vutbr.fit.group.verifit.oslc.compilation.VeriFitCompilationManager;
 import cz.vutbr.fit.group.verifit.oslc.shared.utils.Utils;
 
@@ -40,7 +41,7 @@ public class SutFetchGit implements SutFetcher {
 		        .setDirectory(folderPath.toFile())
 		        .call();
 		    
-		    return Utils.getResourceIdFromUri(new URI (url)); // gets the last part of the URL
+		    return OslcValues.getResourceIdFromUri(new URI (url)); // gets the last part of the URL
 		
 		} catch (GitAPIException | JGitInternalException | URISyntaxException e) {
 			throw new IOException("Git clone failed: " + e.getMessage());