+++ compilation/src/main/java/cz/vutbr/fit/group/verifit/oslc/compilation/automationRequestExecution/sutFetcher/SutFetchUrl.java
@@ -18,6 +18,7 @@ import java.nio.file.Path;
 
 import org.apache.commons.io.FileUtils;
 
+import cz.vutbr.fit.group.verifit.oslc.OslcValues;
 import cz.vutbr.fit.group.verifit.oslc.compilation.VeriFitCompilationManager;
 import cz.vutbr.fit.group.verifit.oslc.shared.utils.Utils;
 
@@ -41,7 +42,7 @@ public class SutFetchUrl implements SutFetcher {
 		}
 		
 		try {
-			String fileName = Utils.getResourceIdFromUri(sourceUrl); // gets the last part of the URL
+			String fileName = OslcValues.getResourceIdFromUri(sourceUrl); // gets the last part of the URL
 			Path pathToFile = folderPath.resolve(fileName);
 			
 			FileUtils.copyURLToFile(